+++ src/opera_utils/tropo/__init__.py
@@ -3,11 +3,19 @@
 from ._apply import apply_tropo
 from ._crop import crop_tropo
 from ._match import apply_tropo_correction, match_and_apply_tropo, read_reference_point
+from ._motion import (
+    estimate_flow,
+    interp_in_time_motion,
+    interpolate_motion,
+)
 
 __all__ = [
     "apply_tropo",
     "apply_tropo_correction",
     "crop_tropo",
+    "estimate_flow",
+    "interp_in_time_motion",
+    "interpolate_motion",
     "match_and_apply_tropo",
     "read_reference_point",
 ]